Create README.md
AWS Parallel Cluster helps in deploying and managing the High-Performance Computing in an automated and secure manner as concurrently as possible and wait until all branches terminate before processing the parallel state. It helps in gaining insights faster and quickly move from an idea to market with virtually unlimited compute capacity resulting high-throughput networking. The message passing interface is a standardized means of computing architectures allows communication of information between various nodes and clusters. As it is one of the most common protocols used in High Performance Computing it helps to provide collective communication involving number of processors at the same time.
The  major aim is to implement an application that can decompose a data set and process it in parallel on a distributed Cloud environment and optimize our solution that can find the best balance of computation, communication and resource costs. The respective data processing will be conducted on a simple matrix operation i.e., Matrix addition and Matrix Multiplication. Implementing the application on AWS cloud platform will provide: -
 Operational Excellence
 Security
 Reliability
 Performance Efficiency
 Sustainability